moving all parsing functions into parse.cpp
